Pointing out where to look for errors
I figure some people won't know to look at the output of the command to check for errors. I haven't tried it on Windows to write an appropriate version, and I'm sure you'd want to tweak the language to match the rest of the documents anyway. You might also want to mention that your sudo password is the same as your Mac password. If